% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Nusselt.R
\name{calc_Nusselt}
\alias{calc_Nusselt}
\title{Full calculation of Nu, h and R_conv}
\usage{
calc_Nusselt(
  Re,
  Pr,
  k,
  Dh,
  A,
  correlation = c("dittus", "gnielinski"),
  heating = TRUE,
  f = NULL
)
}
\arguments{
\item{Re}{Reynolds number}

\item{Pr}{Prandtl number}

\item{k}{Thermal conductivity [W/(m·K)]}

\item{Dh}{Hydraulic diameter [m]}

\item{A}{Heat exchange surface area [m^2]}

\item{correlation}{Correlation to use ("dittus" or "gnielinski")}

\item{heating}{TRUE if wall is heated}

\item{f}{Friction factor (optional for Gnielinski)}
}
\value{
List with Nu, h, R_conv
}
\description{
Full calculation of Nu, h and R_conv
}
