% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_corr .R
\name{compare_corr}
\alias{compare_corr}
\title{Compare Dittus–Boelter and Gnielinski correlations}
\usage{
compare_corr(Re_vec, Pr, k, Dh, A, heating = TRUE)
}
\arguments{
\item{Re_vec}{Vector of Reynolds numbers}

\item{Pr}{Prandtl number}

\item{k}{Thermal conductivity [W/(m·K)]}

\item{Dh}{Hydraulic diameter [m]}

\item{A}{Heat exchange surface area [m^2]}

\item{heating}{TRUE if wall is heated}
}
\value{
List with Nu and h values for both correlations
}
\description{
Compare Dittus–Boelter and Gnielinski correlations
}
