% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu_temperature.R
\name{cpu_temperature}
\alias{cpu_temperature}
\title{Estimate CPU temperature}
\usage{
cpu_temperature(Q, T_amb, R_conv, R_cond = 0.1)
}
\arguments{
\item{Q}{Dissipated power [W]}

\item{T_amb}{Ambient temperature [°C]}

\item{R_conv}{Convective thermal resistance [K/W]}

\item{R_cond}{Conduction thermal resistance [K/W], default 0.1}
}
\value{
CPU temperature [°C]
}
\description{
Estimate CPU temperature
}
