% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_conv.R
\name{h_conv}
\alias{h_conv}
\title{Calculate convection coefficient h}
\usage{
h_conv(Nu, k, Dh)
}
\arguments{
\item{Nu}{Nusselt number}

\item{k}{Thermal conductivity [W/(m·K)]}

\item{Dh}{Hydraulic diameter [m]}
}
\value{
Convection coefficient h [W/(m^2·K)]
}
\description{
Calculate convection coefficient h
}
