% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nusselt_dittus_boelter.R
\name{nusselt_dittus_boelter}
\alias{nusselt_dittus_boelter}
\title{Calculate Nusselt number with Dittus–Boelter correlation}
\usage{
nusselt_dittus_boelter(Re, Pr, heating = TRUE)
}
\arguments{
\item{Re}{Reynolds number}

\item{Pr}{Prandtl number}

\item{heating}{TRUE if wall is heated, FALSE if cooled}
}
\value{
Nusselt number
}
\description{
Calculate Nusselt number with Dittus–Boelter correlation
}
