% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nusselt_gnielinski.R
\name{nusselt_gnielinski}
\alias{nusselt_gnielinski}
\title{Calculate Nusselt number with Gnielinski correlation}
\usage{
nusselt_gnielinski(Re, Pr, f = NULL)
}
\arguments{
\item{Re}{Reynolds number}

\item{Pr}{Prandtl number}

\item{f}{Friction factor (optional). If NULL, smooth pipe approximation is used.}
}
\value{
Nusselt number
}
\description{
Calculate Nusselt number with Gnielinski correlation
}
