% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpu_temp.R
\name{plot_cpu_temp}
\alias{plot_cpu_temp}
\title{Plot CPU temperature vs dissipated power}
\usage{
plot_cpu_temp(
  Q_vec,
  T_amb,
  R_conv,
  R_cond = 0.1,
  main = "CPU Temperature vs Dissipated Power"
)
}
\arguments{
\item{Q_vec}{Vector of dissipated power values [W]}

\item{T_amb}{Ambient temperature [°C]}

\item{R_conv}{Convective thermal resistance [K/W]}

\item{R_cond}{Conduction thermal resistance [K/W], default 0.1}

\item{main}{Plot title}
}
\value{
Data frame with Q and T_cpu
}
\description{
Plot CPU temperature vs dissipated power
}
