% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance_conv.R
\name{resistance_conv}
\alias{resistance_conv}
\title{Calculate convective thermal resistance}
\usage{
resistance_conv(h, A)
}
\arguments{
\item{h}{Convection coefficient [W/(m^2·K)]}

\item{A}{Heat exchange surface area [m^2]}
}
\value{
Convective thermal resistance [K/W]
}
\description{
Calculate convective thermal resistance
}
